lvl 5: shell supports running executable files
The shell supports running executable files as commands within child
processes. If the first piece of the input looks like a path (starts
with . or /) a child process is created and the command specified by
the first argument is executed within the child using the provided
arguments.

Signed-off by: Emily Dror