lvl 6: shell supports running executable in PATH
In the case that the user types something that isn't a path or a
builtin the shell checks whether a file with that name exists in
each of the directories listed in the PATH environment variable in
order. If a file is found the search can stop and that file is
executed with arguments in a child process, else the unrecognized
command error is printed.

Signed-off by: Emily Dror